+++ pyquil/device/__init__.py
@@ -1,4 +1,4 @@
 from pyquil.device._base import AbstractDevice
 from pyquil.device.graph import NxDevice
-from pyquil.device.qcs import QCSDevice
+from pyquil.device.qcs import QCSDevice, get_qcs_device
 from pyquil.device.compiler import CompilerDevice